+++ packages/core/src/config/index.ts
@@ -1,0 +1,2 @@
+export * from "./projectConfig";
+export * from "./parsers";
