+++ packages/core/src/datasource/index.ts
@@ -1,2 +1,3 @@
 export * from "./datasource";
-export * from "./parsers";
+export * from "./adapter";
+export * from "./filesystemAdapter";
